Minor changes + Architectural Fixes
