+++ Contributors_Only/Source/Generators/XORShift.cpp
@@ -1,28 +1,7 @@
-#pragma once
-#include <math.h>
-#include <type_traits>
-
-#ifndef DF_XORSHIFT_H
-#define DF_XORSHIFT_H
-
-#include "generator.h"
+#include "XORShift.h"
 
 namespace DiceForge
-{   
-    /* DiceForge::XORShift - A RNG utilizing Marsgalia's XORShift algorithm 
-    * Supports 32 and 64 bit unsigned integers */
-    template <typename T>
-    class XORShift : public Generator<T>
-    {
-    private:
-        T m_state;
-        T generate() override;
-        void reseed(T seed) override;
-    public:
-        XORShift(T seed);
-        ~XORShift();
-    };
-
+{
     template <typename T>
     XORShift<T>::XORShift(T seed)
         : m_state(seed)
@@ -58,8 +37,4 @@ namespace DiceForge
         return m_state;
     }
 
-    typedef XORShift<uint32_t> XORShift32;
-    typedef XORShift<uint64_t> XORShift64;
 }
-
-#endif
