+++ Contributors_Only/Source/distributions.h
@@ -1,0 +1,8 @@
+#ifndef DF_DISTRIBUTIONS_H
+#define DF_DISTRIBUTIONS_H
+
+#include "distribution.h"
+#include "Continuous/Cauchy.h"
+#include "Continuous/Gibbs.h"
+
+#endif
