+++ Contributors_Only/Source/generators.h
@@ -1,7 +1,7 @@
 #ifndef DF_GENERATORS_H
 #define DF_GENERATORS_H
 
-#include "Generators/xorshift.h"
+#include "Generators/XORShift.h"
 #include "Generators/MT.h"
 #include "Generators/blumblumshub.h"
 #include "Generators/LFSR.h"
