+++ Contributors_Only/Source/types.h
@@ -1,0 +1,16 @@
+#ifndef DF_TYPES_H
+#define DF_TYPES_H
+
+namespace DiceForge
+{
+    typedef double real_t;
+    typedef long int64_t;
+    typedef int int32_t;
+    typedef unsigned int uint32_t;
+    typedef unsigned long uint64_t;
+
+    typedef uint32_t uint_t;
+    typedef int32_t int_t;
+}
+
+#endif
